# chore(react-icons): improve build performance