- moves unfiling functionality to nodejs cli
  - including `--dry-run` mode
  - 100x faster than previous `sed` bash solution
- runs font generation in parallel

## Performance

| task | Before | After |
|--------|--------|--------|
| `react-icons build` | 4m 1s | 2m 27s / 𝚫 39% FASTER |
| `npm run unfill` | 2m | 2s / 𝚫 99% FASTER |

---
Replaces https://github.com/microsoft/fluentui-system-icons/pull/859